Implement Workaround for Lack of Pod Presets on AKS
Pod Presets are an alpha-level feature of Kubernetes, so that
unfortunately means that they're not available yet. They would be an
elegant way to address the need for us to dynamically generate file
share volume mounts without forcing admins to have to modify the
backend deployment config templates directly.

As a workaround, I'm abusing the ability for Bash to store multiple
lines of text in a single variable, as well as the ability for the GNU
`envsubst` to substitute them in. To avoid complaints from editors that
a YAML file is not well-formed, the variable expansion starts on a
comment line, since each expansion is a full fragment of YAML.

Although a little ugly, this still gives admin users the ability to
define file shares in the config file instead of having to hack
templates.

If you want something more elegant, vote for the Pod Presets feature here:
https://feedback.azure.com/forums/914020-azure-kubernetes-service-aks/suggestions/35054089-support-podpreset-alpha-feature