Rename YAML Files to `.template.yaml` to Clarify Intent
For the casual reader, it may not have been obvious that environment
variable expansion within a config file is NOT a feature of Kubernetes.
To aid in understanding, each of the template files that we pre-process
during deployment have been renamed to `.template.yaml` to signify that
they are not usable directly as-is.