Centralize Shared Duplicate Files into Common Folder
This ensures we only have one copy of `entrypoint.sh` to maintain across both the Apache and FPM images. Soon, there will be more files shared between these two.

Since Docker does not allow a `Dockerfile` to reference files outside of the context root, we have to actually have to explicitly specify the Dockerfile to use and then provide a context root of `..` so that the new `nextcloud-common` folder is included for the Docker daemon when assembling the image.