Re-order Dockerfiles to Optimize Transfer
Custom apps are the most likely layer to change frequently, followed by NR. Since the entrypoint and Nextcloud config customizations are part of this resource kit, they're the least-likely to change. So, it makes more sense to put custom apps at the bottom so that layer is easy to swap out without affecting the others. Then, if NR is being updated or toggled on/off, it can also be updated with minimal impact to the Nextcloud core layer.