Switch to Shared Redis Cache for Sessions and Locks
This should help resolve issues with cluster-based deployment (i.e. with a replica count > 1) by ensuring that a user's session is portable across pods.
The Redis cache is automatically created by the deployment; configuration settings for it need to be specified in `config.env` first, though (as with everything else in the AKS kit).

Documentation has been updated to reflect the latest info.