Reduce the Number of Stale Replica Sets
The docs and forums seem to imply that, unless instructed otherwise, Kubernetes keeps all previous replica sets around in case admins need to rollback to a known good configuration. In practice, it looks like AKS caps this to 10 or so previous replicas. For our purposes, two previous replica sets should be fine.