Provide Option to Restart Pods Post-Publish
The default deployment strategy for Kubernetes requires a docker image version bump in order to trigger a rolling deploy. But, before we actually have a v1 release, this requirement makes development cumbersome.

As a workaround, I've added the command-line option `--cycle-pods` to `deploy_nextcloud_app.sh` that implements the recommendation from https://github.com/kubernetes/kubernetes/issues/13488#issuecomment-446101458. This basically "touches" the containers of the existing deployment by setting an unused environment variable to be equal to the current time. That triggers Kubernetes to perform a rolling deployment because it sees that an environment variable has changed.