Move Nextcloud Logs off Azure Files Share
Keeping logs on Azure Files is asking for trouble and likely creating a bottleneck as multiple instances try to write to the same file. Instead, if the log gets moved to ephemeral storage, we can just write it out to standard out within each pod. Later on, something like Elasticsearch can be used to capture and index that output further.