Eliminate Conflicting Redis Config & Version-lock Nextcloud
Using `nextcloud:15-*` lead to the predictable side-effect of automatically -- and unexpectedly -- getting a new version of Nextcloud during an image rebuild. This lead to a conflicting Redis configuration, which does not support a Redis password/key, being added to the deployment. This change removes the conflicting config in favor of our own, then version-locks Nextcloud so we can control version upgrades to prevent this in the future.