Enforce Naming Convention for Services without External IPs
Right now, the convention is that services not directly exposed (i.e. services without an external IP through Azure) have the prefix `internal`. In the case of Nextcloud, though this may seem counter-intuitive because it is exposed through an ingress controller, the service does not directly expose *itself* to the world, so it *is* "internal".

This convention may change in the future, but for now let's keep it consistent.