Fix HTTPS Ingress - Detect Client IP and Frontend Protocol Properly
Since the ingress controller handles HTTPS instead of pushing this concern down to the NGinx instance running Nextcloud. Unfortunately, out-of-the-box the Nextcloud NGinx config wasn't detecting that the ingress controller is using HTTPS, so links and redirects for DAV and other services weren't being routed properly.

Separately, the ingress controller was not passing on the client IP because the default `externalTrafficPolicy` is `Cluster`. For more on this, see this superb blog article by Andrew Sy Kim: https://www.asykim.com/blog/deep-dive-into-kubernetes-external-traffic-policies