Fix "An unknown error has occurred" for Chunked Uploads > 1 MB
This adjusts the ingress controller to be able to take chunks up to 512 MB. Since we're using chunked uploads, Nextcloud usually sends in chunks in 10 MB increments, so there's really no reason to have to force NGinx to have to deal with 1 GB+ file uploads per request. So, I've standardized everything to top out at 512 MB *per request*.

Based on comments in this issue, it appears that Nextcloud has effectively no maximum file size beyond the underlying storage medium and PHP version (32-bit vs 64-bit):
https://github.com/nextcloud/server/issues/8445

Basically, the 512 MB maximum upload setting in "Basic settings" is a red-herring. It appears not to affect anything.