Ensure Context Gets Reset during (Re)Launch of `aks browse`
Calling `az aks get-credentials` has the unfortunate side-effect of resetting the local Kube context to the `default` namepace. This change ensures we switch back after that happens, so that deploying while the dashboard is open doesn't send configurations to the wrong Kube namespace (e.g. if you have `nextcloud-live`, `nextcloud-dev`, etc, instead of running Nextcloud in the `default` namespace).