Pin `latest` Images to Specific Versions + Stop Always Pulling
Best practice is to avoid depending on a`latest` tag since it is difficult to reproduce a specific version or known good state when you encounter issues.
Since we're pinning specific versions, and the version of the images produced and referenced by this project is controlled on CLI by the `CONTAINER_VERSION` environment variable, I see no reason why we should continue using `imagePullPolicy: Always` since it slows down container spin-up time dramatically.