Correct Site Base URL from "/index.php" to "/" on Apache
This ensures that we write-out an `.htaccess` file that includes the necessary rewrite rules for requests against "/x/y/z" to be processed as if they were against "/index.php/x/y/z". The `.htaccess` file is generated by the Nextcloud `occ` CLI utility.

We no longer need to stash the `.htaccess` file away on the file share nor restore it on startup, since we generate it fresh each time the same way the installer would generate it on initial installation.