Workaround for nextcloud-server #17980 / musl #10960
Per suggestions in #musl IRC, trying an approach that uses larger
buffers for directory operations, which is closer to what glibc appears
to be doing as part of readdir(). For reasons unbeknownst to the musl
team, this appears to avoid the paging behavior that causes the defect
to appear.

To implement this approach, a custom build of the musl standard C
library is required. The build happens in a separate build step so that
the final image does not need to contain build tools or the musl source
code.