Do Not Sync Config Directory if Config is Read-only by Flag
Previously, when we were using blob storage for the config, the config
folder would be mounted read-write when the replica count was `1` and
read-only if the replica count was > `1`. Now we use Azure Files so its
always read-write.

Still, it's a bad idea to update the config during startup if there are
multiple containers. This is controlled by the
`NEXTCLOUD_CONFIG_READ_ONLY` environment variable.