Turn Off Redis Session Locking
Regardless of whether this is safe or not, Nextcloud is just way more
stable without session locking. One hypothesis is that a locked session
can cause Nextcloud to fail to load the session, resulting in the
session cookie on the client side being invalidated, causing a cascade
of subsequent request failures.