Mount Azure File Share RO when `NEXTCLOUD_CONFIG_READ_ONLY` is `true`
Not sure why I didn't do this before, but it's essential to ensure the
app stays up and installed anytime that AKS or Azure Files is having a
bad day (e.g. "transport endpoint not connected"). Otherwise, the config
folder goes missing and Nextcloud thinks it needs to be re-installed,
so it overwrites its own config with a blank one.