[ARKNG-175] Re-factor + Support Path Filtering by JWT Claims
Now, all of the logic of our implementation is contained within
`JwtScopedSftpServer`. The server now uses the custom `authorized_paths`
claim in the JWT to control what paths the user is allowed to see. We
create a custom filtered filesystem per session from the JWT claims. If
a JWT is missing this claim, the user is not allowed to authenticate.