[ARKNG-175] Add Support for RSA Keys + Make Secret Fetching Pluggable
This gives us the ability to use an RS256, RS384, or RS512 asymmetric
key instead of just HS256, HS384, or HS512, so that we can avoid a
vulnerability in this application leaking a PSK that could be used to
craft JWTs in the future.

By combining this with pluggable secret fetching, we can accept JWTs
issued by a variety of systems that each have their own key.