[ARKNG-175] Switch to Specifying Allowed RSA Public Keys per Origin
This will allow us to use different JWT signing keys for each system
that's issuing JWTs. So, if one is compromised, the others are safe.

Next, we'll build on this to add restrictions on what origins can access
what paths in the filesystem, so that different origins can only craft
JWTs for access to parts of the filesystem they're supposed to be able
to access.