[ARKNG-175] Make it Possible to Restrict Paths by Origin/Issuer
This makes it possible for the SFTP-WS server configuration to control
which paths a particular origin is able to grant access to via JWT
grants. So, for example, it's possible to configure the server so that
Server A can only grant users access to Path X and Path Y, and Server B
can only grant users access to Path Z.

This is secure because we require each origin to use a specific public
key, thereby eliminating issuer/orgin spoofing. The `Origin` header must
match the "issuer" (`iss` claim) of the JWT, and the JWT must be signed
with a public key that matches the issuer's private key. If a particular
issuer (e.g. Server A) is compromised and its private key stolen, an
attacker can only issue JWTs that grant access to shares that the issuer
normally can access (e.g. Path X and Path Y) but cannot grant issue to
other shares (e.g. Path Z) because they cannot generate a JWT that has a
private key matching an origin that is allowed to access the other
paths.