[ITSA-706] Port SFTP-WS Keypair Generation over to Rigger Sub-command
This allows for overlay-scoped key generation. We do NOT check-in the
generated keys nor provide a way to pack them into encrypted secrets
because there are too many variables with the way an admin can use them
in the origin restrictions file and they are easily regenerated if they
are lost.