[ITSA-706] Delete Old Scripts
The function of almost all of these scripts is now handled by a
combination of Rigger and Kustomize. As for `launch_xdebug_proxy.sh`,
the new Kustomize-based approach makes it much easier to run a local
copy Nextcloud for debugging, so that is preferred to trying to open a
fragile SSH tunnel proxy to a live environment.