[ITSA-706] Re-factor and Clean-up Docker Publishing to Work with Rigger
The new structure does not automatically authenticate with ACR. In
addition, the registry URL and tag are now passed-in by Rigger on the
command line. Other optional features like New Relic are controlled by
environment variables that are provided within each overlay as a
`build.env` file that is now checked-in to source control.

Protection against image overwrites has been added, to avoid
accidentally replacing a versioned tag. This can be overridden