[ITSA-706] Detect Read-only `/var/www/config` as Early as Possible
This does not prevent errors that the config is read-only during
upgrades, but it does ensure that by the time the config is being loaded
the environment is consistent. We want to set this as early as possible
so that any `./occ` commands that need to pick up can do so.