[ITSA-1478] Prefer Burstable Nodes for Burstable Workloads
These pods already had tolerations to be scheduled on burstable nodes,
but Kubernetes would only do that as a last resort. This conveys to
the scheduler that these pods actually prefer that.

Unfortunately, there is not a way to use a taint as a matching
criterion, so we had to duplicate the taint as a label on the node pool.
This was added with a command like the following:

```
az aks nodepool update \
  --resource-group RESOURCE-GROUP \
  --cluster-name CLUSTER-NAME \
  --name NODE-POOL-NAME \
  --labels inveniem.com/workload-type=burstable
```