[ITSA-1478] Run Regular CronJob on Burstable Nodes
This CronJob does light maintenance tasks for Nextcloud every few mins,
so we can schedule it on lower-cost burstable nodes. The other CronJobs,
which include the Nextcloud file scan and the temporary file cleanup,
tend to have continuous file I/O, so they're not likely a good fit for
a burstable host node because they'll exhaust all the I/O credits.