[IT-234] Update MySQL Connection from Environment Variables
Nextcloud generates the config with database settings the first time
that it is run, but from then on it does not automatically update the
variables from the environment. This is a real pain if the password or
hostname of the DB server needs to change, because an admin has to
update the credentials in BOTH the environment AND the config volume,
which is often mounted read-only. So, to support closer to a zero
downtime deployment model for database connection string changes, this
new config snippet should keep the database connection info in sync
with the environment on a read-only config folder, even if what's in the
config file is not in sync.