[IT-120] Apply Latest Upstream Changes for `nginx.conf`
These changes should work all the way up to Nextcloud v28, since that's
the current version and that version is shipping with this config. There
do not appear to be any version-specific config snippets in here.

The changes appear to enhance the cacheability and handling of static
assets, including JavaScript.